More Unit Tests
Implemented some more unit tests to work towards the 0.2.0 release.
